requirements.txt
ImageHash~=4.3.1
requests~=2.32.3
beautifulsoup4~=4.12.3
pillow~=11.0.0
pandas~=2.2.3